requirements.txt
alembic==1.16.5
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
pillow==11.3.0
python-dotenv==1.1.1
reportlab==4.4.4
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
xhtml2pdf==0.2.17
weasyprint==66.0
svglib==1.6.0
pyphen==0.17.2
cssselect2==0.8.0
tinycss2==1.4.0
html5lib==1.1
pypdf==6.1.1
PyYAML==6.0.3
openpyxl==3.1.2
# Si usas MySQL en producción, descomenta la siguiente línea:
# pymysql
